// src/API.ts
export type Restaurant = {
  __typename: "Restaurant",
  id: string,
  name: string,
  description: string,
  city: string,
  createdAt: string,
  updatedAt: string,
};
// ...
export type OnCreateRestaurantSubscription = {
  onCreateRestaurant?:  {
    __typename: "Restaurant",
    id: string,
    name: string,
    description: string,
    city: string,
    createdAt: string,
    updatedAt: string,
  } | null,
};